In RequestContext::setUser() also reset $this->skinName
Viewing a page which transcludes Special:Prefixindex leads to the
default skin being used for the main request, instead of the user's
skin, in a regression of a bug from 2023. Tracing shows that
OutputPage::setupOOUI() calls RequestContext::getSkinName(), and the
skin name is permanently cached.

So, reset the skin name when the user is reset. This is a simple partial
fix.

Bug: T336504
Change-Id: If630f337ba6fb2b889e657d49112b878c3ad5f83
(cherry picked from commit aa01c94b3da722464ab925e2f58f3cd7db5c1107)